+++ src/snyk/common/constants/languageServer.ts
@@ -11,7 +11,6 @@ export const DID_CHANGE_CONFIGURATION_METHOD = 'workspace/didChangeConfiguration
 export const SNYK_HAS_AUTHENTICATED = '$/snyk.hasAuthenticated';
 export const SNYK_ADD_TRUSTED_FOLDERS = '$/snyk.addTrustedFolders';
 export const SNYK_SCAN = '$/snyk.scan';
-export const SNYK_FOLDERCONFIG = '$/snyk.folderConfigs';
 export const SNYK_SCANSUMMARY = '$/snyk.scanSummary';
 export const SNYK_REGISTER_MCP = '$/snyk.registerMcp';
 export const SNYK_TREEVIEW = '$/snyk.treeView';