+++ src/snyk/common/languageServer/receivedFolderConfigsFromLsState.ts
@@ -1,9 +1,0 @@
-let receivedFolderConfigsFromLs = false;
-
-export function getReceivedFolderConfigsFromLs(): boolean {
-  return receivedFolderConfigsFromLs;
-}
-
-export function setReceivedFolderConfigsFromLs(value: boolean): void {
-  receivedFolderConfigsFromLs = value;
-}